requirements.txt
requests
beautifulsoup4
slack_sdk
google-cloud-storage
google-cloud-secret-manager
python-dotenv
